gh-9 added unit tests
